trading_bot/requirements.txt
requests
pyupbit
pandas
numpy
python-dotenv
